+++ src/test/runTest.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 import * as path from 'path'
 import { runTests } from 'vscode-test'
 
@@ -7,9 +8,13 @@ async function main(): Promise<void> {
     // Passed to `--extensionDevelopmentPath`
     const extensionDevelopmentPath = path.resolve(__dirname, '../../')
 
+    console.log('extensionDevelopmentPath', extensionDevelopmentPath)
+
     // The path to test runner
     // Passed to --extensionTestsPath
     const extensionTestsPath = path.resolve(__dirname, './suite/index')
+
+    console.log('extensionTestsPath', extensionTestsPath)
 
     // Download VS Code, unzip it and run the integration test
     await runTests({
@@ -18,7 +23,6 @@ async function main(): Promise<void> {
       launchArgs: ['--disable-extensions'],
     })
   } catch {
-    // eslint-disable-next-line no-console
     console.error('Failed to run tests')
     // eslint-disable-next-line unicorn/no-process-exit
     process.exit(1)